# cmd/apiserver/docs/swagger.yaml
definitions:
  model.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is a sample API with user authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: My API
  version: "1.0"
paths:
  /private/whoami:
    get:
      description: Get information about currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get current user info
      tags:
      - private
  /sessions:
    post:
      consumes:
      - application/json
      description: Authenticate user and create session
      parameters:
      - description: User email
        example: '"user@example.com"'
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: User password
        example: '"securepassword123"'
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid JSON
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create session (login)
      tags:
      - sessions
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with email and password
      parameters:
      - description: User email
        example: '"user@example.com"'
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: User password
        example: '"securepassword123"'
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid JSON
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"